fix:聊天室测试15

+++ src/main/java/com/stgd/voice/controller/AdminUserController.java
@@ -26,6 +26,9 @@ public class AdminUserController {
 
 	@PostMapping("removeAdminUser")
 	public Integer removeAdminUser(@RequestParam Integer id) {
+		if (id == 1){
+			return 0;
+		}
 		return adminUserMapper.deleteById(id);
 	}
